move ocean components config file
